backend/requirements.txt
# --- Core Framework ---
Flask==3.0.0
Werkzeug==3.0.0

# --- Database & ORM ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# --- Authentication & Security ---
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
passlib==1.7.4

# --- API Cross-Origin (CORS) ---
Flask-Cors==4.0.0

# --- Server Production (Wajib untuk Deploy ke Render) ---
gunicorn==21.2.0

# --- Database Driver (Pilih salah satu sesuai database production) ---
# Jika menggunakan PostgreSQL di Render (Direkomendasikan):
psycopg2-binary==2.9.9

# Jika menggunakan MySQL di server lain:
# PyMySQL==1.1.0

# --- Environment Variables Management ---
python-dotenv==1.0.0